CNV-5024: Inform for pending changes in VM details view
Add alerts when the user make changes in the VM details view
while the VM is running informing to restart the VM for
the changes to apply.

Additionally, the changes aren't reflected immediately on the screen
while the VM is running, like they used to be, but change after the
VM restarts. This way the screen represents the current configuration
rather than the 'future' configuration.

Signed-off-by: Ido Rosenzwig <irosenzw@redhat.com>

CNV-5024# Please enter the commit message for your changes. Lines starting